update submodule commit reference
learning to use a submodule: you pull changes to the submodule
from within its directory, treating it as its own repository.
Then when you pop up out of that directory and git diff, the
parent repo notices that the head is at a different commit than
it was before. You can then add and commit that difference.